Instead of doing one seek for every 30 second chunk, do a single seek to read the whole clip.
This requires a substantial rearchitecture, plus some buffering for the sake of speed, but
the result is faster and dramatically more accurate.  In particular, the errors caused by
inaccurate seeking were responsible for the apparent clock drift between tracks.  With this
method, no false drift appears, so no drift correction is needed.