Use audiorate to clean up the input sample stream.
This fixes a problem observed with an ogg file that was missing a page
(and thus was partially corrupt).  The missing page created a desync
between the streaming time and the seeking time.  With this change,
that file can be aligned correctly, and we have no known failed test
cases.

(The ogg corruption may be due to #629196, which has since been fixed.)